BaseUser entity added. Database issues fixed

+++ src/main/java/com/cies/cies_project/model/repositories/DepartmentManagerRepository.java
@@ -3,5 +3,5 @@ package com.cies.cies_project.model.repositories;
 import com.cies.cies_project.model.entities.DepartmentManager;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepartmentManagerRepository extends JpaRepository<DepartmentManager,Integer> {
+public interface DepartmentManagerRepository extends JpaRepository<DepartmentManager, Long> {
 }
